Fixed and added all the necessary fixes so the app works. Added security and login. Added an add chore button and some users. Next need to add the delete button, the add to ChoreAssignment button since we now only have chores and a user, and the map for entering the address.

+++ Choremonkeys/src/main/java/com/example/choremonkeys/services/ChoreService.java
@@ -2,8 +2,6 @@ package com.example.choremonkeys.services;
 
 import com.example.choremonkeys.models.Chore;
 import com.example.choremonkeys.models.ChoreStatus;
-import com.example.choremonkeys.models.User;
-import com.example.choremonkeys.models.UserType;
 
 import java.util.List;
 
@@ -17,14 +15,8 @@ public interface ChoreService {
     List<Chore> findByStatus(ChoreStatus choreStatus);
     List<Chore> findByPriceRange(Long minPrice, Long maxPrice);
 
-    // CREATE operations
 
-    Chore createChoreWithStatus(String title,
-                                String description,
-                                String destination,
-                                Integer phoneNumber,
-                                Long price,
-                                ChoreStatus choreStatus);
+    Chore createChoreWithStatus(String title, String description, String destination, Long phoneNumber, int price, ChoreStatus choreStatus);
 
     // UPDATE operations
     Chore updateChore(Long id,
